backend/requirements.txt
Django==4.2.24
djangorestframework==3.15.2
django-cors-headers==4.3.1
django_extensions==4.1.0